requirements.txt
fastapi
pydantic
python-dotenv
pymongo
passlib[bcrypt]
python-jose
bson
python-multipart
httpx